gauguin: Use CodeAurora ImsService implementation for RCS
There is no evidence that Qualcomm devices running a rather recent IMS
stack need to include AOSP RCS packages. RCS already worked before this
change and continues to do so, but the main goal was to get rid of some
of the RCS related log spam, which was accomplished here.

This reverts commit 918bef58194492f2e39fc9c331b24a5a7d20d3c4.

Change-Id: I219546e29d288373877e9bcd194a60ef5e09f224